Check potential fix: construct communicator in test directly, rather than through fixture
